\name{LIVEdata}
\alias{LIVEdata}
\alias{livedata}
\title{The data generating for the LIVE}
\description{The data generating for the LIVE.
  \itemize{
    \item Columns 1: The response variable y.
    \item Columns 2: The design matrix x.
    \item Columns 3-52: TheThe instrumental variables matrix z.
    }
}
\usage{data(LIVEdata)}
\keyword{datasets}

