\name{naivedata}
\alias{naivedata}
\title{The data generating for the naivereg}
\description{The z and residuals are both generated from a normal distribution.

x = sin(z1)+exp(z5)+z4^2+z15+log(z19+8)+e1.

y = 2*x+e2.
  \itemize{
    \item Columns 1: The design matrix x.
    \item Columns 2: The response variable y.
    \item Columns 3-22: TheThe instrumental variables matrix z.
    }
}
\usage{data(naivedata)}
\keyword{datasets}

