\name{naivedata}
\alias{naivedata}
\title{The data generating for the naivereg}
\description{The data generating for the naivereg.
  \itemize{
    \item Columns 1: The design matrix x.
    \item Columns 2: The response variable y.
    \item Columns 3-102: TheThe instrumental variables matrix z.
    }
}
\usage{data(naivedata)}
\keyword{datasets}

