\name{naivedata}
\docType{data}
\alias{naivedata}
\title{Data generated by Monte Carlo}
\description{
Data generated by Monte Carlo, the first column is the response variable, the second column is the design matrix, and the rest are the instrumental variables
}
\usage{data(naivedata)}


\keyword{datasets}
